requirements.txt
pandas
numpy
sqlalchemy
psycopg2
psycopg2-binary
mysql-connector-python